add support for 404 page; refactor homepage render based on cookie value
