+++ src/App.js
@@ -4,9 +4,6 @@ import React, {useState, useEffect} from 'react';
 /* Import React and framer motion components */
 import {Helmet} from "react-helmet";
 
-/* Import cookies */
-import { useCookies } from 'react-cookie';
-
 /* Import assets */
 import './assets/css/main.css';
 
@@ -17,76 +14,28 @@ import { PRELOADIMAGES } from "./components/PreloadImages";
 import MainNav from "./components/MainNav";
 import AnimatedRoutes from "./components/AnimatedRoutes";
 
-import gsap from 'gsap';
-import {useGSAP} from '@gsap/react';
-import {ScrollTrigger} from "gsap/ScrollTrigger";
-gsap.registerPlugin(ScrollTrigger);
-
 function App() {
   const [imgsLoaded, setImgsLoaded] = useState(false); // get images for preload
-  const [entImageIndex, setEntImageIndex] = useState(1); // set image iterator
-  const [animAppearance, setAnimAppearance] = useState("block"); // hide animation after completion
-  const [visitCookie, setVisitCookie] = useCookies(false); // set cookie to detect if scroll animation already ran
-  const [scroll, setScroll] = useState(false); // check idf user has scrolled
-
+  
   useEffect(() => {
     // Begin preload function
-    const loadImage = image => {
+    const loadImage = (image) => {
       return new Promise((resolve, reject) => {
         const loadImg = new Image();
         loadImg.src = image.url;
-        // wait 1 second to simulate loading time
-        loadImg.onload = () =>
+        loadImg.onload = () => resolve(image.url);
           setTimeout(() => {
-            resolve(image.url)
-          }, 1000);
+            resolve(image.url);
+          }, 0);
 
-        loadImg.onerror = err => reject(err)
+        loadImg.onerror = err => reject(err);
       })
     }
 
     Promise.all(PRELOADIMAGES.map(image => loadImage(image)))
       .then(() => setImgsLoaded(true))
       .catch(err => console.log("Failed to load images", err))
-
-    window.addEventListener("scroll", () => {
-      setScroll(true);
-    });
-
   }, []);
-
-  const body = document.querySelector('body');
-
-    useGSAP(() => {
-      const stop = 21; // number of images to display on scroll
-      var i = 1;
-  
-      ScrollTrigger.create({
-        markers: false,
-        trigger: body,
-        start: 'top',
-        onUpdate: self => {
-          if (self.direction === 1) {
-            i = i + 1;
-            setEntImageIndex(i);
-            if (i === stop) {
-              setAnimAppearance("none");
-              setVisitCookie('scrollRan', true);
-            }
-          } else {
-            i = i - 1;
-            setEntImageIndex(i);
-          }
-        },
-        onLeaveBack: () => {
-          setEntImageIndex(1);
-        },
-      });
-
-      const tl = gsap.timeline();
-      tl.to('#ScrollPrompt', {repeat: -1, y: -15, yoyo: true});
-    });
-
 
   return (
     <>
@@ -103,32 +52,14 @@ function App() {
     </Helmet>
     <main>
         {imgsLoaded ? 
-          <> 
-          { visitCookie.scrollRan ? 
-            <>
-              <header>
-                <MainNav />
-                <hr className="header-border" />
-              </header>
-              <AnimatedRoutes />
-            </>
-            :
-            <>
-              <div className="draw-jacket">
-                <aside id="ScrollPrompt" role="alert" className={scroll ? "scroll-prompt-fade" : ""}>
-                  <p>Scroll down<br />to draw me</p>
-                </aside>
-                <img 
-                  style={{display: animAppearance}} 
-                  id="ScrollIntroAnim" 
-                  alt="on scroll animation intro" 
-                  src={"./anim-images/intro/drawintro_" + entImageIndex + ".svg"} 
-                />
-              </div>
-            </>
-          }
+          <>
+            <header>
+              <MainNav />
+              <hr className="header-border" />
+            </header>
+            <AnimatedRoutes />
           </>
-         : 
+          :
           <> 
             <h1 id="loadingTitle">Loading portfolio</h1>
           </>
